switched to new UI architecture: from UIViewController to UITableViewController + added pull to refresh feature + added view controller for audio player
