+++ VKAudioPlayer/AudioPlayerViewController.swift
@@ -1,0 +1,35 @@
+//
+//  AudioPlayerViewController.swift
+//  VKAudioPlayer
+//
+//  Created by Nikita Belousov on 5/28/16.
+//  Copyright © 2016 Nikita Belousov. All rights reserved.
+//
+
+import UIKit
+
+class AudioPlayerViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
