+++ VKAudioPlayer/MasterViewController+SearchDelegate.swift
@@ -1,5 +1,5 @@
 //
-//  MasterViewController+SearchDelegate.swift
+//  TableViewController+SearchDelegate.swift
 //  VKAudioPlayer
 //
 //  Created by Nikita Belousov on 5/24/16.
@@ -9,14 +9,24 @@
 import UIKit
 import VK_ios_sdk
 
-extension MasterViewController: UISearchResultsUpdating, UISearchBarDelegate {
+extension TableViewController: UISearchResultsUpdating, UISearchBarDelegate {
     
     func updateSearchResultsForSearchController(searchController: UISearchController) {
+        
+        if !searchController.active {
+            refreshControl = UIRefreshControl()
+            refreshControl!.addTarget(self, action: #selector(refresh(_:)), forControlEvents: UIControlEvents.ValueChanged)
+        }
+        
         if searchController.active {
+            refreshControl?.endRefreshing()
+            refreshControl = nil
+            
             context.cancel()
             let audioRequestDescription = AudioRequestDescription.searchAudioRequestDescription(searchController.searchBar.text!)
             initializeContext(audioRequestDescription)
         }
+        
     }
     
     func searchBarCancelButtonClicked(searchBar: UISearchBar) {
